# Use a faster `deflate` setting