@jsgf 2016-11-30
BTW, there's a zstd binding crate on crates.io.